Move AssertionResults to common directory
