+++ Source/Testably.Expectations/Core/Results/AssertionResultWhich.cs
@@ -4,7 +4,7 @@ using System.Linq.Expressions;
 using System.Runtime.CompilerServices;
 using Testably.Expectations.Core.Helpers;
 
-namespace Testably.Expectations.Core;
+namespace Testably.Expectations.Core.Results;
 
 /// <summary>
 ///     The result of an assertion with an underlying value of type <typeparamref name="TResult" />.
