+++ Source/Testably.Expectations/Delegates/ThatDelegate.WithValue.cs
@@ -1,5 +1,6 @@
 ﻿using Testably.Expectations.Core;
 using Testably.Expectations.Core.Helpers;
+using Testably.Expectations.Core.Results;
 
 namespace Testably.Expectations.Expectations;
