+++ Source/Testably.Expectations/Delegates/ThatDelegate.WithoutValue.cs
@@ -1,5 +1,6 @@
 ﻿using Testably.Expectations.Core;
 using Testably.Expectations.Core.Helpers;
+using Testably.Expectations.Core.Results;
 using Testably.Expectations.Core.Sources;
 
 namespace Testably.Expectations.Expectations;
