+++ Source/Testably.Expectations/Primitives/ThatExceptionExtensions.cs
@@ -3,6 +3,7 @@ using System.Runtime.CompilerServices;
 using Testably.Expectations.Core;
 using Testably.Expectations.Core.Formatting;
 using Testably.Expectations.Core.Helpers;
+using Testably.Expectations.Core.Results;
 using Testably.Expectations.Core.Sources;
 
 // ReSharper disable once CheckNamespace
